# packages/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Persona API",
    "summary": "Hosted service for building and running typed-memory AI personas.",
    "version": "0.8.0"
  },
  "paths": {
...
    "/v1/me/usage": {
      "get": {
        "tags": ["me"],
        "summary": "Get Usage",
        "description": "The caller's per-turn token usage (\u00a75.5; turn_logs, RLS-scoped).",
        "operationId": "get_usage_v1_me_usage_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UsageEntry"
                  },
                  "title": "Response Get Usage V1 Me Usage Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/profile": {
      "get": {
        "tags": ["me"],
        "summary": "Get Profile",
        "description": "The caller's own profile \u2014 identity + optional name (Spec K6, K6-D-1).\n\nNull-safe: ``first_name``/``last_name`` are ``None`` for a nameless account.\nThe row is provisioned by ``ensure_user`` in the auth dependency, so a 404 here\nmeans a genuine invariant break rather than a first-time user.",
        "operationId": "get_profile_v1_me_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["me"],
        "summary": "Update Profile",
        "description": "Set the caller's optional name + timezone (Spec K6/A8). PATCH \u2014 omitted = unchanged.\n\nOnly the fields the client actually sent are written (``exclude_unset``): a\nstring sets, an explicit ``null`` clears, an omitted field is left untouched.\nNames are normalised (control-char strip, whitespace-only \u2192 unset) in the\nservice (K6-D-8). A provided ``timezone`` is validated as an IANA zone here\n(Spec A8, A8-D-9) \u2014 an unknown zone is a fail-fast 422, never stored to\nmis-fire in the tick; ``null`` clears it (\u2192 falls back to the config default).\nScoped to the caller's own row \u2014 never another user's.",
        "operationId": "update_profile_v1_me_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/schedule/occurrences": {
      "get": {
        "tags": ["me"],
        "summary": "Get Schedule Occurrences",
        "description": "The caller's upcoming schedule occurrences + fire history (Spec A8, A8-D-11).\n\nComputed from the engine's own recurrence path (never a client reimplementation), RLS-scoped\nto the caller. The window is server-capped (horizon + count); the response's ``truncated``\nmarker says so honestly when a wide ``from/to`` is clamped. ``from`` must be ``<= to``.",
        "operationId": "get_schedule_occurrences_v1_me_schedule_occurrences_get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OccurrencesResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/schedule/{schedule_id}/reschedule/preview": {
      "post": {
        "tags": ["me"],
        "summary": "Preview Schedule Reschedule",
        "description": "Preview a calendar reschedule \u2014 the engine's next-fire + full clause + quiet-hours warn.\n\nNo write (Spec A8, T9, bars 3/5): the picker shows this as the confirm echo (the SAME full\nclause chat re-echoes) before the user confirms. The next fire is the ENGINE's, so the picker\nnever fabricates a time the DST gap/fold policy would shift.",
        "operationId": "preview_schedule_reschedule_v1_me_schedule__schedule_id__reschedule_preview_post",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Schedule Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleRescheduleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReschedulePreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/schedule/{schedule_id}/reschedule": {
      "post": {
        "tags": ["me"],
        "summary": "Apply Schedule Reschedule",
        "description": "Apply a calendar reschedule through the SAME CAS door as chat (Spec A8, T9, bar 4).\n\n``actor=user_via_ui``; the client sends picker-state (no raw RRULE \u2014 the server maps it). RLS-\nscoped to the caller. Returns the applied clause + the engine's next fire (the twin's confirm).",
        "operationId": "apply_schedule_reschedule_v1_me_schedule__schedule_id__reschedule_post",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Schedule Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleRescheduleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReschedulePreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/notifications": {
      "get": {
        "tags": ["me"],
        "summary": "Get Notifications",
        "description": "The caller's durable bell feed (Spec P6, RLS-scoped, newest-first, paginated).",
        "operationId": "get_notifications_v1_me_notifications_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationOut"
                  },
                  "title": "Response Get Notifications V1 Me Notifications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/me/notifications/{notification_id}/read": {
      "post": {
        "tags": ["me"],
        "summary": "Mark Notification Read",
        "description": "Mark one notification read (deep-link click, Spec P6). 0 if not owned/absent.",
        "operationId": "mark_notification_read_v1_me_notifications__notification_id__read_post",
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationMarkReadResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FireEvent": {
        "properties": {
          "schedule_id": {
            "type": "string",
            "title": "Schedule Id"
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["schedule_id", "at", "status"],
        "title": "FireEvent",
        "description": "One past fire/miss from the audit trail (the calendar's ran/missed markers)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotificationMarkReadResult": {
        "properties": {
          "updated": {
            "type": "integer",
            "title": "Updated"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["updated"],
        "title": "NotificationMarkReadResult",
        "description": "How many feed rows a mark-read touched (0 = nothing unread / not owned)."
      },
...
      "OccurrencesResult": {
        "properties": {
          "occurrences": {
            "items": {
              "$ref": "#/components/schemas/Occurrence"
            },
            "type": "array",
            "title": "Occurrences"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/FireEvent"
            },
            "type": "array",
            "title": "History"
          },
          "window_from": {
            "type": "string",
            "format": "date-time",
            "title": "Window From"
          },
          "window_to": {
            "type": "string",
            "format": "date-time",
            "title": "Window To"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "occurrences",
          "history",
          "window_from",
          "window_to",
          "truncated"
        ],
        "title": "OccurrencesResult",
        "description": "The windowed occurrences + fire history + the honest truncation marker (A8-D-11)."
      },
...
      "ReschedulePreview": {
        "properties": {
          "human_terms": {
            "type": "string",
            "title": "Human Terms"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "next_fire": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Fire"
          },
          "quiet_hours_offer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quiet Hours Offer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "human_terms",
          "timezone",
          "next_fire",
          "quiet_hours_offer"
        ],
        "title": "ReschedulePreview",
        "description": "The engine's preview of a proposed calendar reschedule (no write)."
      },
...
      "ScheduleRescheduleRequest": {
        "properties": {
          "pattern": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecurrencePattern"
              },
              {
                "type": "null"
              }
            ]
          },
          "one_time_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "One Time At"
          },
          "timezone": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Timezone"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["timezone"],
        "title": "ScheduleRescheduleRequest",
        "description": "A calendar-initiated reschedule (Spec A8, T9 \u2014 the twin of the chat verb).\n\n**Picker-state in \u2014 NO raw RRULE from the client** (bar 1): the web sends a structured\n:class:`~persona.schedules.RecurrencePattern` (or a one-time instant) and the SERVER maps it to\nthe rule, so there is no client-side recurrence math. Exactly one of ``pattern`` /\n``one_time_at`` (XOR, checked in the route). Applied through the SAME CAS door as chat."
      },
...
      "UpdateProfileRequest": {
        "properties": {
          "first_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "quiet_hours_start": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1439.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Quiet Hours Start"
          },
          "quiet_hours_end": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1439.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Quiet Hours End"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UpdateProfileRequest",
        "description": "Set the caller's optional name + timezone (Spec K6/A8). PATCH semantics.\n\nAll fields optional; **omitted = unchanged**, explicit ``null`` = **clear**\n(distinguished server-side via ``model_dump(exclude_unset=True)``). ``max_length``\nfails fast at the boundary on egregious input; the service then strips control\ncharacters and treats whitespace-only as unset (names,\n:func:`persona_api.services.user_service.normalize_name`). ``timezone`` (Spec A8,\nA8-D-9) is an IANA zone name whose validity is checked in the route handler\n(:func:`persona.timezone.validate_timezone` \u2192 422) \u2014 the ``max_length`` here is\nonly a cheap egregious-input guard, not the IANA check (a custom ``field_validator``\nwould break the 422 body's ``json.dumps``, cf. ``SendMessageRequest``). ``null``\nclears it \u2192 schedule computation falls back to ``PERSONA_DEFAULT_TIMEZONE``.\nNothing is required \u2014 an empty PATCH is a valid no-op read."
      },
      "UsageEntry": {
        "properties": {
          "persona_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Id"
          },
          "tier_used": {
            "type": "string",
            "title": "Tier Used"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "prompt_tokens": {
            "type": "integer",
            "title": "Prompt Tokens"
          },
          "completion_tokens": {
            "type": "integer",
            "title": "Completion Tokens"
          },
          "cost_cents": {
            "type": "number",
            "title": "Cost Cents"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tier_used",
          "model_name",
          "prompt_tokens",
          "completion_tokens",
          "cost_cents",
          "created_at"
        ],
        "title": "UsageEntry",
        "description": "One usage-log row (per-turn telemetry, paginated)."
      },
      "UserProfileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "first_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "quiet_hours_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quiet Hours Start"
          },
          "quiet_hours_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quiet Hours End"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["id", "created_at"],
        "title": "UserProfileResponse",
        "description": "The caller's own profile \u2014 identity anchor + optional name (Spec K6).\n\n``first_name`` / ``last_name`` are optional (``None`` when unset \u2014 a nameless\naccount is fully valid). ``email`` may be ``None`` when the token carried none\n(the provisioning fallback stores a noreply address, but the surface stays\nnullable so the contract never implies a real address). Our DB is the source of\ntruth; Clerk stays auth-only."
      },
...
    }
  }
}